Convert to float, not int

+++ lib/proforma/prawn_renderer/balanced_set.rb
@@ -16,7 +16,7 @@ module Proforma
           free_spots = values.count(&:nil?)
 
           if free_spots.positive?
-            available_value = sum - values.map(&:to_i).reduce(0, :+)
+            available_value = sum - values.map(&:to_f).reduce(0, :+)
 
             new_values = divide(free_spots, available_value, round)
 
@@ -29,7 +29,7 @@ module Proforma
         private
 
         def balance(values, sum, round)
-          difference = sum - values.map(&:to_i).reduce(0, :+)
+          difference = sum - values.map(&:to_f).reduce(0, :+)
 
           return values if difference.zero?
